add DocItem custom props to Link element
This minor change would allow custom props to be leveraged in CSS, typically using `::before`or `::after`pseudo-elements to inject content with `attr()`. The custom props themselves could be specified manually or injected by a custom sidebar items generator. Currently, one has to swizzle the component in order to take advantage of custom props, which is overkill in a number of cases.